app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
python-dotenv
python-multipart
aiofiles
